requirements.txt
azure-core==1.35.0
azure-storage-blob==12.26.0
boto3==1.40.2
botocore==1.40.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.5
idna==3.10
isodate==0.7.2
jmespath==1.0.1
pycparser==2.22
pyotp==2.9.0
pyPreservica==3.2.3
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
